nix/logrotate: Don't copy file but use contents right away.
While copying seems okay to do, here's what happens:  `File` reads the contents of the source file in the configure phase (if possible) and computes the target state. `File` then writes the contents in the update phase. So the NixOS component -- in its configure phase -- reads the file on disk from the last batou run, because the parent component has (of course) not yet written the new file. Tada: old file being copied.