[FC-35973] systemd timers non-persistent by default
Systemd Timers were persistent by default previously.
Since that implies a trigger on first activation, this is unwanted
behaviour for Jobs that interact with the outside world, e.g. sending
out emails.

The old behaviour can be enabled optionally.
Looking over some uses of this Component in our deployments,
I did not find any that rely on the old behaviour or ones
that would not work with the new default.