batou_ext.python: add FixELFRunPath
Alternative to using LD_LIBRARY_PATH for Python libraries with
precompiled .so libraries which has caused unintended side-effects
(FC-35412, FC-36567).

The proper, long-term solution is of course to use Nix for that, however
that will be a long road. Until then, this tooling may help to fix up
the paths where C libraries are loaded from without tainting the
environment.