[FC-38393] Fix container verification not triggering correctly
This PR fixes a case where the update is not triggered even though
the running and locally available running and the locally available
image are different, because the digest has been has been added
to the cache.

I also restructured the code a bit. By using early returns
and thereby removing deeply nested structures, the control flow
becomes much easier to follow