batou_ext.oci: Add support for `podman`
FC-37959

We mostly want this for healthchecks that must pass before the unit is
actually active.

This is kept intentionally simple, only healthchecks via sd_notify are
supported. Mixing containers in systemd units with `healthy` and `conmon`
strategy leads to setups with conflicting requirements (linger vs no
linger) and it's far too easy to end up with an unsupported
configuration (and containers not behaving properly).