[FC-42763] fix multiple trash dirs on one vm
Adjust the merging behaviour of the `serviceConfig` option value by
defining the values as lists so that multiple trash directories can be defined per VM.

Previously, multiple definitions would collide when merging since they differ.
Defining the value as a list makes the option's values mergeable in the
indended way.