ssh: support multiple ssh keys per deployment
FC-44942

Right now, it's possible to only add two keys per deployment (with the
attributes id_ed25519 and id_rsa). Now that we have more applications
that we pull from GitHub which only allows a key to be used once, we
need a way to enroll multiple SSH keys in a deployment.

This is done via `file_name`. When used, it's possible to add either a
id_ed25519 or a id_rsa, but not both.