[FC-49329] fix SymlinkandCleanup for directories
Due to how the `dirname` and `basename` functions work, they made using
this component with folders entirely impossible.
Using the very mature `pathlib` library solves this with the `parent`
and `parts` attributes.

This change also has the nice side-effect of not requiring absolute
paths to be passed as `current` since `dirname("foo.txt")` is an
empty string and `os.chdir("")` (from further down) fails with an
exception.
`Path("foo.txt").parent` on the other hand returns `"."`
which does not break `os.chdir`.